+++ sdk/cosmos/azure-cosmos/azure/cosmos/_cosmos_client_connection.py
@@ -2049,6 +2049,7 @@ class CosmosClientConnection:  # pylint: disable=too-many-public-methods,too-man
         # Patch will use WriteEndpoint since it uses PUT operation
         request_params = RequestObject(resource_type, documents._OperationType.Patch)
         request_params.set_excluded_location_from_options(options)
+        request_params.set_retry_write(options.pop('retryWrite', False))
         request_data = {}
         if options.get("filterPredicate"):
             request_data["condition"] = options.get("filterPredicate")
@@ -2655,6 +2656,7 @@ class CosmosClientConnection:  # pylint: disable=too-many-public-methods,too-man
 
         request_params = RequestObject(typ, documents._OperationType.Create)
         request_params.set_excluded_location_from_options(options)
+        request_params.set_retry_write(options.pop('retryWrite', False))
         result, last_response_headers = self.__Post(path, request_params, body, headers, **kwargs)
         self.last_response_headers = last_response_headers
 
@@ -2702,6 +2704,7 @@ class CosmosClientConnection:  # pylint: disable=too-many-public-methods,too-man
         # Upsert will use WriteEndpoint since it uses POST operation
         request_params = RequestObject(typ, documents._OperationType.Upsert)
         request_params.set_excluded_location_from_options(options)
+        request_params.set_retry_write(options.pop('retryWrite', False))
         result, last_response_headers = self.__Post(path, request_params, body, headers, **kwargs)
         self.last_response_headers = last_response_headers
         # update session for write request
@@ -2746,6 +2749,7 @@ class CosmosClientConnection:  # pylint: disable=too-many-public-methods,too-man
         # Replace will use WriteEndpoint since it uses PUT operation
         request_params = RequestObject(typ, documents._OperationType.Replace)
         request_params.set_excluded_location_from_options(options)
+        request_params.set_retry_write(options.pop('retryWrite', False))
         result, last_response_headers = self.__Put(path, request_params, resource, headers, **kwargs)
         self.last_response_headers = last_response_headers
 
@@ -2828,6 +2832,7 @@ class CosmosClientConnection:  # pylint: disable=too-many-public-methods,too-man
         # Delete will use WriteEndpoint since it uses DELETE operation
         request_params = RequestObject(typ, documents._OperationType.Delete)
         request_params.set_excluded_location_from_options(options)
+        request_params.set_retry_write(options.pop('retryWrite', False))
         result, last_response_headers = self.__Delete(path, request_params, headers, **kwargs)
         self.last_response_headers = last_response_headers